requirements.txt
langchain==0.2.6
langchain-google-genai==1.0.6
langchain-community==0.2.6
langchainhub==0.1.20
langchain-experimental==0.0.62
streamlit==1.36.0
python-dotenv==1.0.1
duckduckgo-search==6.1.8
tavily-python==0.3.3
beautifulsoup4==4.12.3
faiss-cpu==1.8.0.post1
numpy==1.26.4
pypdf==4.2.0
pandas==2.2.2
tabulate==0.9.0
seaborn==0.13.2
matplotlib==3.8.4
# --- Pinning core dependencies to fix gRPC/Auth errors ---
google-api-python-client==2.134.0
google-auth-httplib2==0.2.0
google-auth==2.29.0
grpcio==1.64.1
protobuf==4.25.3
